fix more LP64 signed-shift UB found by UBSan live test
- misc.c get32(): uint32 accumulator (already committed separately)
- ftpserv.c pport(): uint32 n for FTP PORT/EPRT address assembly
- sntp.c dosntpsysrefid(): uint32 refid accumulator (stratum 1 refid)
- timer.c/hpux.c: Msclock = 1000 * Secclock overflows int32 for any
  epoch > 1970-04-26; compute in 64-bit, cast on assignment
- iproute.c/ipfilter.c: ~0L << (32-bits) is 64-bit-wide mask on LP64,
  masking wrong bits; use ~0U (32-bit) - a behavior bug, not just UB

+++ src/ftpserv.c
@@ -683,7 +683,7 @@ static
 int
 pport(struct socket *sock,char *arg)
 {
-	int32 n;
+	uint32 n;
 	int i;
 
 	n = 0;
@@ -693,7 +693,7 @@ pport(struct socket *sock,char *arg)
 			return -1;
 		arg++;
 	}
-	sock->address = n;
+	sock->address = (int32) n;
 	n = atoi(arg);
 	if((arg = strchr(arg,',')) == NULL)
 		return -1;
